Fix CCA detection in Cooja in the case when channels are switched on receiver during a transmission.
The patch always adds a connection on transmission, even when the channel is not correct. Initially the connection is in a dormant state, similarly to when the receiver is turned off;
when the receiver is turned on and switched to the right channel, `updateSignalStrengths()` is called, and the connection starts to recieve PHY-level traffic.